Committing personal keymap to VCS
I'll be honest, I had a fork that I made when I was still learning git,
the thing was a poor mess. I tried to clean it up and pull in changes
from upstream, but I'm still learning git (lol).  So here I am, forking
again just to have a sensible history.